fixed REALLY annoying collapsing sidebar bug
The bug occured if you had the sidebar toggled open (aka #wrapper was toggled) and you stretched the screen past the medium screen width breakpoint (min-width: 768px) the sidebar would collapse. Then you had a wide screen with no sidebar and the only way to get it back was to reduce the screen to less than 768px and un-toggle the sidebar.