Prevent user from spamming submit for reviews/multiple reviews for same class
If required fields are filled, removes submit button so that user can't spam click it or click it multiple times. Avoids multiple submissions of the same review.

Also created backend validation when creating a new review so that a user doesn't create a review for the same course under the same instructor/semester again. Instead gives them an error message to edit their original review.