Prevent user from spamming submit for reviews/multiple reviews for sa… (#399)
* Prevent user from spamming submit for reviews/multiple reviews for same class

If required fields are filled, removes submit button so that user can't spam click it or click it multiple times. Avoids multiple submissions of the same review.

Also created backend validation when creating a new review so that a user doesn't create a review for the same course under the same instructor/semester again. Instead gives them an error message to edit their original review.

* Made timeout 3 seconds to ensure

* Refactored some of the code

* Fix pylint issue

* Tell user if they have already reviewed the course

Not finished as I am having trouble debugging but would like to get some sleep now O.O

* User notified of duplicate review and unable to spam click submit button upon submission

Seems to work as intended now but I think it's too late at this point to rush it into the upcoming patch. Some more testing is required and tweaks as deemed necessary.

* Resolve nagging pylint error

* Made modifications per discussion with @alex-shen1

Namely, updated the modal to match site theme, modified the filter to check if the user has reviewed for the given course with the same instructor OR semester before, and updated some formatting.